# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "岚山区劳动争议调解区块链平台，后台采用golang开发，使用gin + gorm + gorbac + jwt开发",
        "title": "岚山区劳动争议调解区块链平台",
        "termsOfService": "http://hafrans.com",
        "contact": {
            "name": "Chuuka Ro (Hafrans)",
            "url": "http://hafrans.com/support",
            "email": "lvzh@hafrans.com"
        },
        "license": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/meeting/account/create": {
            "get": {
                "description": "创建会议专用账户，只有管理员和部门人员可以创建",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meeting"
                ],
                "summary": "创建会议专用账户",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vo.CommonData"
                        }
                    },
                    "401": {
                        "description": "未验证",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
        "/api/v1/meeting/account/delete": {
            "get": {
                "description": "删除会议专用账户，只有管理员和部门人员可以操作",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meeting"
                ],
                "summary": "删除会议专用账户",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vo.CommonData"
                        }
                    },
                    "401": {
                        "description": "未验证",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
        "/api/v1/meeting/cancel/:id": {
            "get": {
                "description": "取消会议，只有管理员和创建会议者可以取消",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meeting"
                ],
                "summary": "取消会议",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vo.CommonData"
                        }
                    },
                    "401": {
                        "description": "未验证",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
        "/api/v1/meeting/create": {
            "post": {
                "description": "创建会议，只有管理员和部门人员可以创建",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meeting"
                ],
                "summary": "创建会议",
                "parameters": [
                    {
                        "description": "新建会议表单",
                        "name": "meeting",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.MeetingCreateForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vo.CommonData"
                        }
                    },
                    "401": {
                        "description": "未验证",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
        "/api/v1/meeting/list": {
            "get": {
                "description": "列出相关会议，管理员可以看到全部的，部门管理员只可以看到自己的",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meeting"
                ],
                "summary": "列出相关会议",
                "parameters": [
                    {
                        "type": "number",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "页大小",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vo.CommonData"
                        }
                    },
                    "401": {
                        "description": "未验证",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "vo.Common": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "unauthorized"
                },
                "status": {
                    "type": "integer",
                    "example": 401
                },
                "timestamp": {
                    "type": "string",
                    "example": "2048-05-06 12:34:56"
                }
            }
        },
        "vo.CommonData": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "unauthorized"
                },
                "status": {
                    "type": "integer",
                    "example": 401
                },
                "timestamp": {
                    "type": "string",
                    "example": "2048-05-06 12:34:56"
                }
            }
        },
...
        "vo.MeetingCreateForm": {
            "type": "object",
            "required": [
                "case_id",
                "end_time",
                "meeting_type",
                "start_time",
                "subject"
            ],
            "properties": {
                "allow_unmute_self": {
                    "description": "允许参会者取消静音",
                    "type": "boolean"
                },
                "case_id": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "hosts": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "invitee": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "meeting_type": {
                    "description": "会议类型",
                    "type": "integer"
                },
                "mute_all": {
                    "description": "全体静音",
                    "type": "boolean"
                },
                "mute_enable_join": {
                    "description": "入会时静音",
                    "type": "boolean"
                },
                "password": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                }
            }
        },
...
    }
}